move hydrateSchema to its own file
